pyproject.toml
[project]
name = "mcp-sse"
version = "0.1.0"
description = "SSE-based MCP Server and Client"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.45.1",
    "argparse>=1.4.0",
    "httpx>=0.28.1",
    "litellm>=1.0.0",
    "mcp[cli]>=1.2.1",
    "python-dotenv>=1.0.1",
]
